@@ -27,11 +27,11 @@ project_urls =
   Source = https://github.com/sappelhoff/pyprep
 
 [options]
-python_requires = >= 3.7
+python_requires = >= 3.8
 install_requires =
-  numpy >= 1.14.1
-  scipy >= 1.0.0
-  mne >= 0.20.0
+  numpy >= 1.20.2
+  scipy >= 1.6.3
+  mne >= 1.3.0
   psutil >= 5.4.3
 packages = find:
 include_package_data = True
